Extract observer module
